requirements.txt
Cython==0.29.21
h5py==2.10.0
imageio==2.6.1
ipython==7.9.0
ipython-genutils==0.2.0
matplotlib==3.1.1
nibabel==2.5.1
numpy==1.18.1
odl==1.0.0.dev0
opencv-python==4.1.1.26
Pillow==6.0.0
pyparsing==2.4.2
python-dateutil==2.8.0
requests==2.18.4
scikit-image==0.16.2
scikit-learn==0.21.3
scipy==1.2.1
SimpleITK==1.2.3
six==1.14.0
tensorboard==1.13.1
tensorboardX==1.8
tensorflow==1.14.0
tensorflow-estimator==1.13.0
tensorflow-gpu==1.4.0
tensorflow-tensorboard==0.4.0
torch==1.4.0
torchfile==0.1.0
torchstat==0.0.7
torchsummary==1.5.1
torchvision==0.5.0
tqdm==4.36.1

